@MrCroxx 2024-05-15
<img width="679" alt="image" src="https://github.com/MrCroxx/foyer/assets/22407295/345744c2-08c3-469b-ab0b-5babd863d17b">
<img width="1274" alt="image" src="https://github.com/MrCroxx/foyer/assets/22407295/104b92c8-49db-4e2c-88c7-a930fb4dcbb3">

Looks good. 🥵