@MrCroxx 2024-05-15
## Benchmark

### PR

```bash
rm -rf /p44pro/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=warn ./target/release/foyer-bench --dir /p44pro/foyer --mem 64 --disk 102400 --file-size 64 --get-range 10000 --flushers 12 --reclaimers 12 --time 60 --writers 64 --w-rate 8 --admission-rate-limit 500 --readers 1024 --r-rate 1 --runtime --compression none --distribution zipf --distribution-zipf-s 0.8
```

```
Total:
disk total iops: 6775.8
disk total throughput: 467.8 MiB/s
disk read iops: 131.9
disk read throughput: 8.8 MiB/s
disk write iops: 6643.9
disk write throughput: 459.0 MiB/s
insert iops: 6706.9/s
insert throughput: 419.2 MiB/s
insert lat p50: 3us
insert lat p90: 5us
insert lat p99: 8us
insert lat p999: 65us
insert lat p9999: 146us
insert lat p99999: 211us
insert lat pmax: 291us
get iops: 15848.1/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 114us
get hit lat p90: 347us
get hit lat p99: 607us
get hit lat p999: 5151us
get hit lat p9999: 6943us
get hit lat p99999: 7839us
get hit lat pmax: 7871us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```

<img width="1470" alt="image" src="https://github.com/MrCroxx/foyer/assets/22407295/5fee3329-2ab4-4e74-82e3-606dd5c6fd4c">

### main

```bash
rm -rf /p44pro/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=warn ./target/release/foyer-storage-bench --dir /p44pro/foyer --capacity 102400 --file-size 64 --get-range 10000 --flushers 12 --reclaimers 12 --time 60 --writers 64 --w-rate 8 --ticket-insert-rate-limit 500 --readers 1024 --r-rate 1 --runtime --metrics --compression none --distribution zipf --distribution-zipf-s 0.8
```

```
Total:
disk total iops: 7735.0
disk total throughput: 513.0 MiB/s
disk read iops: 824.1
disk read throughput: 54.7 MiB/s
disk write iops: 6910.8
disk write throughput: 458.3 MiB/s
insert iops: 6736.2/s
insert throughput: 421.0 MiB/s
insert lat p50: 37us
insert lat p90: 71us
insert lat p99: 393us
insert lat p999: 931us
insert lat p9999: 1743us
insert lat p99999: 2559us
insert lat pmax: 3023us
get iops: 16233.1/s
get miss: 2.37%
get throughput: 1.0 GiB/s
get hit lat p50: 899us
get hit lat p90: 2271us
get hit lat p99: 3007us
get hit lat p999: 4415us
get hit lat p9999: 7519us
get hit lat p99999: 9023us
get hit lat pmax: 9471us
get miss lat p50: 144us
get miss lat p90: 315us
get miss lat p99: 647us
get miss lat p999: 1175us
get miss lat p9999: 2143us
get miss lat p99999: 2623us
get miss lat pmax: 2623us
```

<img width="1470" alt="image" src="https://github.com/MrCroxx/foyer/assets/22407295/abe8f5b4-4f1f-4135-b24f-5f72ac674aea">